Run refactor

+++ rts/src/adjustor.rs
@@ -1,7 +1,7 @@
 use crate::prelude::*;
 
-unsafe fn totalArgumentSize(mut typeString: *const c_char) -> c_int {
-    let mut sz = 0 as c_int;
+unsafe fn totalArgumentSize(mut typeString: *const c_char) -> i32 {
+    let mut sz = 0;
 
     while *typeString != 0 {
         let fresh0 = typeString;
@@ -10,10 +10,10 @@ unsafe fn totalArgumentSize(mut typeString: *const c_char) -> c_int {
         let mut t = *fresh0;
         let mut current_block_1: u64;
 
-        match t as c_int {
+        match t as i32 {
             100 | 108 | 76 => {
-                if size_of::<*mut c_void>() as usize == 4 as usize {
-                    sz += 2 as c_int;
+                if size_of::<*mut c_void>() as usize == 4 {
+                    sz += 2;
                     current_block_1 = 4906268039856690917;
                 } else {
                     current_block_1 = 12333325718155969978;
@@ -26,7 +26,7 @@ unsafe fn totalArgumentSize(mut typeString: *const c_char) -> c_int {
 
         match current_block_1 {
             12333325718155969978 => {
-                sz += 1 as c_int;
+                sz += 1;
             }
             _ => {}
         }
