+++ rts/src/ghcautoconf.rs
@@ -1,9 +1,9 @@
 use crate::prelude::*;
 
-pub(crate) const RTS_LINKER_USE_MMAP: c_int = 1 as c_int;
+pub(crate) const RTS_LINKER_USE_MMAP: i32 = 1;
 
 #[ffi(ghc_lib)]
-pub const SIZEOF_LONG: c_int = 8 as c_int;
+pub const SIZEOF_LONG: i32 = 8;
 
 #[ffi(ghc_lib)]
-pub const SIZEOF_VOID_P: c_int = 8 as c_int;
+pub const SIZEOF_VOID_P: i32 = 8;
