+++ rts/src/posix/get_env.rs
@@ -4,10 +4,10 @@ unsafe fn get_environ() -> *mut *mut c_char {
     return *_NSGetEnviron();
 }
 
-unsafe fn getProgEnvv(mut out_envc: *mut c_int, mut out_envv: *mut *mut *mut c_char) {
-    let mut envc: c_int = 0;
+unsafe fn getProgEnvv(mut out_envc: *mut i32, mut out_envv: *mut *mut *mut c_char) {
+    let mut envc: i32 = 0;
     let mut environ = get_environ();
-    envc = 0 as c_int;
+    envc = 0;
 
     while !(*environ.offset(envc as isize)).is_null() {
         envc += 1;
@@ -17,4 +17,4 @@ unsafe fn getProgEnvv(mut out_envc: *mut c_int, mut out_envv: *mut *mut *mut c_c
     *out_envv = environ;
 }
 
-unsafe fn freeProgEnvv(mut envc: c_int, mut envv: *mut *mut c_char) {}
+unsafe fn freeProgEnvv(mut envc: i32, mut envv: *mut *mut c_char) {}
