+++ rts/src/testsuite/testmblockalloc.rs
@@ -3,45 +3,45 @@ use crate::ffi::rts::storage::m_block::{freeMBlocks, getMBlocks, releaseFreeMemo
 use crate::ffi::rts_api::{RtsOptsAll, defaultRtsConfig, hs_init_ghc};
 use crate::prelude::*;
 
-static mut MAXALLOC: c_int = 16 as c_int;
+static mut MAXALLOC: i32 = 16;
 
-static mut ARRSIZE: c_int = 64 as c_int;
+static mut ARRSIZE: i32 = 64;
 
-static mut LOOPS: c_int = 1000 as c_int;
+static mut LOOPS: i32 = 1000;
 
-static mut SEED: c_int = 0xf00f00 as c_int;
+static mut SEED: i32 = 0xf00f00;
 
-unsafe fn main_0(mut argc: c_int, mut argv: *mut *mut c_char) -> c_int {
-    let mut i: c_int = 0;
-    let mut j: c_int = 0;
-    let mut b: c_int = 0;
+unsafe fn main_0(mut argc: i32, mut argv: *mut *mut c_char) -> i32 {
+    let mut i: i32 = 0;
+    let mut j: i32 = 0;
+    let mut b: i32 = 0;
     let vla = ARRSIZE as usize;
     let mut a: Vec<*mut c_void> = ::std::vec::from_elem(null_mut::<c_void>(), vla);
     let vla_0 = ARRSIZE as usize;
-    let mut sizes: Vec<uint32_t> = ::std::vec::from_elem(0, vla_0);
-    srand(SEED as c_uint);
+    let mut sizes: Vec<u32> = ::std::vec::from_elem(0, vla_0);
+    srand(SEED as u32);
 
     let mut conf = defaultRtsConfig;
     conf.rts_opts_enabled = RtsOptsAll;
     hs_init_ghc(&raw mut argc, &raw mut argv, conf);
-    i = 0 as c_int;
+    i = 0;
 
     while i < LOOPS {
-        j = 0 as c_int;
+        j = 0;
 
         while j < ARRSIZE {
-            if i > 0 as c_int {
+            if i > 0 {
                 freeMBlocks(
                     *a.as_mut_ptr().offset(j as isize),
                     *sizes.as_mut_ptr().offset(j as isize),
                 );
             }
 
-            b = rand() % MAXALLOC + 1 as c_int;
+            b = rand() % MAXALLOC + 1;
 
             let ref mut fresh5 = *a.as_mut_ptr().offset(j as isize);
-            *fresh5 = getMBlocks(b as uint32_t);
-            *sizes.as_mut_ptr().offset(j as isize) = b as uint32_t;
+            *fresh5 = getMBlocks(b as u32);
+            *sizes.as_mut_ptr().offset(j as isize) = b as u32;
             j += 1;
         }
 
@@ -49,7 +49,7 @@ unsafe fn main_0(mut argc: c_int, mut argv: *mut *mut c_char) -> c_int {
     }
 
     releaseFreeMemory();
-    j = 0 as c_int;
+    j = 0;
 
     while j < ARRSIZE {
         freeMBlocks(
@@ -61,23 +61,23 @@ unsafe fn main_0(mut argc: c_int, mut argv: *mut *mut c_char) -> c_int {
     }
 
     releaseFreeMemory();
-    i = 0 as c_int;
+    i = 0;
 
     while i < LOOPS {
-        j = 0 as c_int;
+        j = 0;
 
         while j < ARRSIZE {
-            b = rand() % MAXALLOC + 1 as c_int;
+            b = rand() % MAXALLOC + 1;
 
             let ref mut fresh6 = *a.as_mut_ptr().offset(j as isize);
-            *fresh6 = getMBlocks(b as uint32_t);
-            *sizes.as_mut_ptr().offset(j as isize) = b as uint32_t;
+            *fresh6 = getMBlocks(b as u32);
+            *sizes.as_mut_ptr().offset(j as isize) = b as u32;
             j += 1;
         }
 
-        j = ARRSIZE - 1 as c_int;
+        j = ARRSIZE - 1;
 
-        while j >= 0 as c_int {
+        while j >= 0 {
             freeMBlocks(
                 *a.as_mut_ptr().offset(j as isize),
                 *sizes.as_mut_ptr().offset(j as isize),
@@ -91,7 +91,7 @@ unsafe fn main_0(mut argc: c_int, mut argv: *mut *mut c_char) -> c_int {
 
     releaseFreeMemory();
     hs_exit();
-    exit(0 as c_int);
+    exit(0);
 }
 
 fn main() {
@@ -111,7 +111,7 @@ fn main() {
 
     unsafe {
         ::std::process::exit(main_0(
-            (args_ptrs.len() - 1) as c_int,
+            (args_ptrs.len() - 1) as i32,
             args_ptrs.as_mut_ptr() as *mut *mut c_char,
         ) as i32)
     }
